Bug 661658. Fix pointer as int size problem on Win64. r=joedrew
The glyph cache uses a pointer for a hash value, and uses
that hash value as a pointer. We need to ensure that the
hash type is large enough to fit a pointer.

UltraBlame original commit: 17f2489e16604e9b21e31734f6b534f05ba2b84f

+++ gfx/cairo/cairo/src/cairo-cache-private.h
@@ -84,7 +84,7 @@
 
 
 typedef struct _cairo_cache_entry {
-    unsigned long hash;
+    uintptr_t hash;
     unsigned long size;
 } cairo_cache_entry_t;
